perfume_pricing_app_pro/requirements.txt
pandas>=2.1.0
openpyxl>=3.1.2
rapidfuzz>=3.9.0
streamlit>=1.36.0
python-dateutil>=2.8.2
pyyaml>=6.0.2
